Move add mention to filter
