+++ snowflake/filters/__init__.py
@@ -1,2 +1,3 @@
 from .iso_time import iso_time
 from .humanize_time import humanize_time
+from .mention import add_mentions
